requirements.txt
youtube-dl==2020.1.1
discord.py==1.2.5
async-timeout==3.0.1
urllib3==1.25.7
python-dotenv==0.10.3
PyNaCl==1.3.0
praw==6.4.0
prawcore==1.0.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
googletrans==2.4.0
wolframalpha==3.0.1